Move back the permission handler in menu activity
